+++ apps/android_deploy/app/src/main/jni/tvm_runtime.h
@@ -27,12 +27,12 @@
 #include "../src/runtime/c_runtime_api.cc"
 #include "../src/runtime/cpu_device_api.cc"
 #include "../src/runtime/workspace_pool.cc"
-#include "../src/runtime/module_util.cc"
-#include "../src/runtime/system_lib_module.cc"
+#include "../src/runtime/library_module.cc"
+#include "../src/runtime/system_library.cc"
 #include "../src/runtime/module.cc"
 #include "../src/runtime/registry.cc"
 #include "../src/runtime/file_util.cc"
-#include "../src/runtime/dso_module.cc"
+#include "../src/runtime/dso_library.cc"
 #include "../src/runtime/thread_pool.cc"
 #include "../src/runtime/object.cc"
 #include "../src/runtime/threading_backend.cc"
