+++ apps/android_rpc/app/src/main/jni/tvm_runtime.h
@@ -39,12 +39,12 @@
 #include "../src/runtime/c_runtime_api.cc"
 #include "../src/runtime/cpu_device_api.cc"
 #include "../src/runtime/workspace_pool.cc"
-#include "../src/runtime/module_util.cc"
-#include "../src/runtime/system_lib_module.cc"
+#include "../src/runtime/library_module.cc"
+#include "../src/runtime/system_library.cc"
 #include "../src/runtime/module.cc"
 #include "../src/runtime/registry.cc"
 #include "../src/runtime/file_util.cc"
-#include "../src/runtime/dso_module.cc"
+#include "../src/runtime/dso_library.cc"
 #include "../src/runtime/rpc/rpc_session.cc"
 #include "../src/runtime/rpc/rpc_event_impl.cc"
 #include "../src/runtime/rpc/rpc_server_env.cc"
