+++ apps/howto_deploy/tvm_runtime_pack.cc
@@ -40,7 +40,7 @@
 #include "../../src/runtime/c_runtime_api.cc"
 #include "../../src/runtime/cpu_device_api.cc"
 #include "../../src/runtime/workspace_pool.cc"
-#include "../../src/runtime/module_util.cc"
+#include "../../src/runtime/library_module.cc"
 #include "../../src/runtime/module.cc"
 #include "../../src/runtime/registry.cc"
 #include "../../src/runtime/file_util.cc"
@@ -55,8 +55,8 @@
 // Likely we only need to enable one of the following
 // If you use Module::Load, use dso_module
 // For system packed library, use system_lib_module
-#include "../../src/runtime/dso_module.cc"
-#include "../../src/runtime/system_lib_module.cc"
+#include "../../src/runtime/dso_library.cc"
+#include "../../src/runtime/system_library.cc"
 
 // Graph runtime
 #include "../../src/runtime/graph/graph_runtime.cc"
