+++ src/runtime/dso_library.cc
@@ -18,14 +18,14 @@
  */
 
 /*!
- * \file dso_module.cc
- * \brief Module to load from dynamic shared library.
+ * \file dso_libary.cc
+ * \brief Create library module to load from dynamic shared library.
  */
 #include <tvm/runtime/module.h>
 #include <tvm/runtime/memory.h>
 #include <tvm/runtime/registry.h>
 #include <tvm/runtime/packed_func.h>
-#include "module_util.h"
+#include "library_module.h"
 
 #if defined(_WIN32)
 #include <windows.h>
@@ -36,51 +36,19 @@
 namespace tvm {
 namespace runtime {
 
-// Module to load from dynamic shared libary.
+// Dynamic shared libary.
 // This is the default module TVM used for host-side AOT
-class DSOModuleNode final : public ModuleNode {
+class DSOLibrary final : public Library {
  public:
-  ~DSOModuleNode() {
+  ~DSOLibrary() {
     if (lib_handle_) Unload();
   }
-
-  const char* type_key() const final {
-    return "dso";
-  }
-
-  PackedFunc GetFunction(
-      const std::string& name,
-      const ObjectPtr<Object>& sptr_to_self) final {
-    BackendPackedCFunc faddr;
-    if (name == runtime::symbol::tvm_module_main) {
-      const char* entry_name = reinterpret_cast<const char*>(
-          GetSymbol(runtime::symbol::tvm_module_main));
-      CHECK(entry_name!= nullptr)
-          << "Symbol " << runtime::symbol::tvm_module_main << " is not presented";
-      faddr = reinterpret_cast<BackendPackedCFunc>(GetSymbol(entry_name));
-    } else {
-      faddr = reinterpret_cast<BackendPackedCFunc>(GetSymbol(name.c_str()));
-    }
-    if (faddr == nullptr) return PackedFunc();
-    return WrapPackedFunc(faddr, sptr_to_self);
-  }
-
   void Init(const std::string& name) {
     Load(name);
-    if (auto *ctx_addr =
-        reinterpret_cast<void**>(GetSymbol(runtime::symbol::tvm_module_ctx))) {
-      *ctx_addr = this;
-    }
-    InitContextFunctions([this](const char* fname) {
-        return GetSymbol(fname);
-      });
-    // Load the imported modules
-    const char* dev_mblob =
-        reinterpret_cast<const char*>(
-            GetSymbol(runtime::symbol::tvm_dev_mblob));
-    if (dev_mblob != nullptr) {
-      ImportModuleBlob(dev_mblob, &imports_);
-    }
+  }
+
+  void* GetSymbol(const char* name) final {
+    return GetSymbol_(name);
   }
 
  private:
@@ -88,6 +56,12 @@ class DSOModuleNode final : public ModuleNode {
 #if defined(_WIN32)
   // library handle
   HMODULE lib_handle_{nullptr};
+
+  void* GetSymbol_(const char* name) {
+    return reinterpret_cast<void*>(
+        GetProcAddress(lib_handle_, (LPCSTR)name)); // NOLINT(*)
+  }
+
   // Load the library
   void Load(const std::string& name) {
     // use wstring version that is needed by LLVM.
@@ -96,12 +70,10 @@ class DSOModuleNode final : public ModuleNode {
     CHECK(lib_handle_ != nullptr)
         << "Failed to load dynamic shared library " << name;
   }
-  void* GetSymbol(const char* name) {
-    return reinterpret_cast<void*>(
-        GetProcAddress(lib_handle_, (LPCSTR)name)); // NOLINT(*)
-  }
+
   void Unload() {
     FreeLibrary(lib_handle_);
+    lib_handle_ = nullptr;
   }
 #else
   // Library handle
@@ -113,20 +85,23 @@ class DSOModuleNode final : public ModuleNode {
         << "Failed to load dynamic shared library " << name
         << " " << dlerror();
   }
-  void* GetSymbol(const char* name) {
+
+  void* GetSymbol_(const char* name) {
     return dlsym(lib_handle_, name);
   }
+
   void Unload() {
     dlclose(lib_handle_);
+    lib_handle_ = nullptr;
   }
 #endif
 };
 
 TVM_REGISTER_GLOBAL("module.loadfile_so")
 .set_body([](TVMArgs args, TVMRetValue* rv) {
-    auto n = make_object<DSOModuleNode>();
+    auto n = make_object<DSOLibrary>();
     n->Init(args[0]);
-    *rv = runtime::Module(n);
+    *rv = CreateModuleFromLibrary(n);
   });
 }  // namespace runtime
 }  // namespace tvm
